pyproject.toml
[tool.poetry]
name = "numba-raytracer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "numba_raytracer"}]

[tool.poetry.dependencies]
python = "^3.11"
numba = "^0.57.1"
pillow = "^10.0.0"
typer = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
